Implemented trades and a matching engine to handle incoming trades

+++ include/order.h
@@ -2,14 +2,14 @@
 
 #include <chrono>
 #include <string>
+#include <cstdint>
 
 
 // OrderId is a type alias for unsigned int
-using OrderID = unsigned int;
+using OrderID = uint64_t;
 
 // Side represents which direction an order is: buying or selling
 enum class Side { BUY, SELL };
-
 
 // Order<PriceT, QtyT> is the abstract base for all order types
 template<typename PriceT, typename QtyT>
@@ -29,7 +29,6 @@ class Order {
     // Without, deleting a LimitOrder* through an Order* would be undefined behavior
     virtual ~Order() = default;
 
-
     // Const getter methods
     OrderID id() const { return id_; }
     Side side() const { return side_; }
@@ -37,12 +36,11 @@ class Order {
     PriceT price() const { return price_; }
     std::chrono::time_point<std::chrono::steady_clock> timestamp() const { return timestamp_; }
 
-
     // Pure virtual methods: every concrete order type must answer these
     // is_marketable(): can this order match immediately against the book
     // type_str(): human-readable name for printing
     virtual bool is_marketable() const = 0;
-    virtual std::string type_str() const = 0;
+    virtual std::string type_str() const = 0;  // NOTE: might be worth changing to a pure virtual enum for strictness
 
  private: 
     // All fields are const. Once an order is created, nothing changes.
